# backend/movieframe/api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8443
  description: Generated server url
paths:
  /api/users/{userName}:
    put:
      tags:
      - user-rest-controller
      summary: Update user
      operationId: movieUpdating
      parameters:
      - name: userName
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "400":
          description: Invalid id supplied
        "404":
          description: No movie with this id was found to update
  /api/movies/{id}:
    get:
      tags:
      - movie-rest-controller
      summary: Get movie
      operationId: getMovieAPI
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Invalid id supplied
        "200":
          description: Found movie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movie'
        "404":
          description: No movie with this id was found
    put:
      tags:
      - movie-rest-controller
      summary: Update movie
      operationId: movieUpdating_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Movie'
        required: true
      responses:
        "200":
          description: Updated movie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movie'
        "400":
          description: Invalid id supplied
        "404":
          description: No movie with this id was found to update
  /api/users:
    post:
      tags:
      - user-rest-controller
      summary: Post user
      operationId: newUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "200":
          description: Posted user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "404":
          description: No user posted
  /api/users/{id}/image:
    post:
      tags:
      - user-rest-controller
      summary: Post user image
      operationId: uploadUserImageAPI
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              required:
              - imageFile
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
      responses:
        "404":
          description: No movie with this id was found to post it an image
        "400":
          description: Invalid id supplied
        "200":
          description: Posted user image
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
  /users/{userName}/image:
    get:
      tags:
      - user-rest-controller
      operationId: downloadUserImage
      parameters:
      - name: userName
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /reviews:
    get:
      tags:
      - review-rest-controller
      operationId: getReviews
      parameters:
      - name: page
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageReview'
...
  /api/users/{userName}/image:
    get:
      tags:
      - user-rest-controller
      summary: Get user image
      operationId: downloadUserImageAPI
      parameters:
      - name: userName
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: No movie image with this id was found
        "400":
          description: Invalid id supplied
        "200":
          description: Found user image
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /api/users/current:
    get:
      tags:
      - user-rest-controller
      summary: Get current user
      operationId: getCurrentUser
      responses:
        "200":
          description: Found users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "404":
          description: No user founded
...
components:
  schemas:
    Review:
      type: object
      properties:
        id:
          type: integer
          format: int64
        rating:
          type: integer
          format: int32
        coments:
          type: string
        coment:
          type: string
          writeOnly: true
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        encodedPassword:
          type: string
        name:
          type: string
        email:
          type: string
        roles:
          type: string
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/Review'
        imageFile:
          type: object
          properties:
            binaryStream:
              type: object
        review:
          $ref: '#/components/schemas/Review'
    Director:
      type: object
      properties:
        id:
          type: integer
          format: int64
        director:
          type: string
        biography:
          type: string
        name:
          type: string
        born:
          type: string
        residence:
          type: string
        score:
          type: number
          format: double
        imageFile:
          type: object
          properties:
            binaryStream:
              type: object
        genre:
          type: array
          items:
            type: string
    Movie:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        category:
          type: string
        image:
          type: boolean
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/Review'
        directors:
          type: array
          items:
            $ref: '#/components/schemas/Director'
        description:
          type: string
        imageFile:
          type: object
          properties:
            binaryStream:
              type: object
        review:
          $ref: '#/components/schemas/Review'
        votes:
          type: integer
          format: int32
...
    Pageable:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/Sort'
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        paged:
          type: boolean
        unpaged:
          type: boolean
    Sort:
      type: object
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean
    PageReview:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/Review'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/Sort'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/Pageable'
        empty:
          type: boolean
...